backend/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.1"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asyncpg==0.30.0",
    "authx==1.4.2",
    "bcrypt==4.3.0",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "click==8.1.8",
    "cryptography==44.0.2",
    "dnspython==2.7.0",
    "ecdsa==0.19.1",
    "email-validator==2.2.0",
    "fastapi==0.115.12",
    "fastapi-cli==0.0.7",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "implicit>=0.7.2",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "joblib==1.4.2",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "numpy==1.26.4",
    "pandas==2.2.3",
    "pyasn1==0.4.8",
    "pycparser==2.22",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pydantic-settings==2.8.1",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-jose==3.4.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "rich==14.0.0",
    "rich-toolkit==0.14.1",
    "rsa==4.9",
    "scikit-learn==1.6.1",
    "scikit-surprise==1.1.4",
    "scipy==1.15.2",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "starlette==0.46.1",
    "threadpoolctl==3.6.0",
    "typer==0.15.2",
    "typing-extensions==4.13.1",
    "typing-inspection==0.4.0",
    "tzdata==2025.2",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "watchfiles==1.0.5",
    "websockets==15.0.1",
]
